port extensions to the user and org models
