+++ lib/travis/scheduler/models/organization.rb
@@ -1,0 +1,7 @@
+Organization.class_eval do
+  has_one :subscription, as: :owner
+
+  def subscribed?
+    subscription.present? and subscription.active?
+  end
+end
